requirements_all.txt
websockets>=12.0
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
python-dotenv>=1.0
anyio>=3.6
pydantic>=2.6
websockets
python-dotenv
jinja2
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.21.0